Revert "Api data service (#3)"
This reverts commit ff9464ecd86cbe99090c437038795b89936bfadf.
