--- src/main/java/com/example/anton_stock_feed/exceptions/APIServiceException.java
@@ -1,8 +1,0 @@
-package com.example.anton_stock_feed.exceptions;
-
-public class APIServiceException extends RuntimeException {
-
-    public APIServiceException(Throwable cause) {
-        super(cause);
-    }
-}
