+++ src/main/java/com/example/anton_stock_feed/exceptions/CompanyProfileException.java
@@ -1,6 +1,8 @@
 package com.example.anton_stock_feed.exceptions;
 
-public class CompanyProfileException extends RuntimeException {
+import com.example.anton_stock_feed.servlets.CompanyProfile;
+
+public class CompanyProfileException extends RuntimeException{
 
     public CompanyProfileException(Throwable cause) {
         super(cause);
