+++ src/main/java/com/example/anton_stock_feed/exceptions/DAOException.java
@@ -1,8 +1,8 @@
 package com.example.anton_stock_feed.exceptions;
 
-public class DAOException extends RuntimeException {
+public class DAOException extends RuntimeException{
 
-    public DAOException(Throwable cause) {
+    public DAOException (Throwable cause) {
         super(cause);
     }
 }
