+++ src/main/java/com/example/anton_stock_feed/service/CompanyProfileDAO.java
@@ -1,10 +1,8 @@
-package com.example.anton_stock_feed.dao;
+package com.example.anton_stock_feed.service;
 
 import com.example.anton_stock_feed.model.Company;
 
 public interface CompanyProfileDAO {
 
     Company getInfo(String companySymbol);
-
-    void writeData(Iterable<Company> companies);
 }
