+++ src/main/java/com/example/anton_stock_feed/service/CompanyProfileServiceFactory.java
@@ -1,15 +1,13 @@
 package com.example.anton_stock_feed.service;
 
-import com.example.anton_stock_feed.dao.CompanyProfileDAO;
-
 public class CompanyProfileServiceFactory {
 
     public CompanyProfileService createCompanyProfileService(String type, CompanyProfileDAO companyProfileDAO) {
         CompanyProfileService companyProfileService = null;
         if (type.equals("Mock")) {
-            companyProfileService = CompanyProfileServiceMock.getInstance(companyProfileDAO);
+            companyProfileService = new CompanyProfileServiceMock(companyProfileDAO);
         } else if (type.equals("Database")) {
-            companyProfileService = CompanyProfileServiceDatabase.getInstance(companyProfileDAO);
+            companyProfileService = new CompanyProfileServiceDatabase(companyProfileDAO);
         }
         return companyProfileService;
     }
