--- src/main/java/com/example/anton_stock_feed/service/JsonSerialize.java
@@ -1,11 +1,0 @@
-package com.example.anton_stock_feed.service;
-
-import java.util.Collection;
-
-public interface JsonSerialize {
-    String serialize(Object company);
-
-    String serialize(Collection companies);
-
-    Object deserialize(String json);
-}
