--- src/main/java/com/example/anton_stock_feed/service/JsonSerializeFactory.java
@@ -1,11 +1,0 @@
-package com.example.anton_stock_feed.service;
-
-public class JsonSerializeFactory {
-    public JsonSerialize createJsonSerialize(String type) {
-        JsonSerialize jsonSerialize = null;
-        if (type.equals("Gson")) {
-            jsonSerialize = JsonSerializeGson.getInstance();
-        }
-        return jsonSerialize;
-    }
-}
