+++ src/main/java/com/example/anton_stock_feed/servlets/CompanyProfile.java
@@ -1,7 +1,7 @@
 package com.example.anton_stock_feed.servlets;
 
-import com.example.anton_stock_feed.dao.CompanyProfileDAO;
-import com.example.anton_stock_feed.dao.CompanyProfileDAOFactory;
+import com.example.anton_stock_feed.service.CompanyProfileDAO;
+import com.example.anton_stock_feed.service.CompanyProfileDAOFactory;
 import com.example.anton_stock_feed.service.CompanyProfileService;
 import com.example.anton_stock_feed.service.CompanyProfileServiceFactory;
 import com.example.anton_stock_feed.model.Company;
@@ -12,7 +12,7 @@ import javax.servlet.annotation.*;
 import java.io.IOException;
 import java.io.PrintWriter;
 
-@WebServlet(name = "company", value = "/company")
+@WebServlet(name = "CompanyProfile", value = "/CompanyProfile")
 public class CompanyProfile extends HttpServlet {
     CompanyProfileServiceFactory companyProfileServiceFactory;
     CompanyProfileService companyProfileService;
@@ -46,7 +46,7 @@ public class CompanyProfile extends HttpServlet {
         }
         PrintWriter out = response.getWriter();
 
-        if (companySymbolFromURL.equals("company")) {
+        if (companySymbolFromURL.equals("CompanyProfile")) {
             response.setStatus(404);
             response.getWriter().println("No such company");
         } else {
