+++ src/main/webapp/index.jsp
@@ -1,0 +1,13 @@
+<%@ page contentType="text/html; charset=UTF-8" pageEncoding="UTF-8" %>
+<!DOCTYPE html>
+<html>
+<head>
+    <title>JSP - Hello World</title>
+</head>
+<body>
+<h1><%= "Stock feed" %>
+</h1>
+<br/>
+<a href="CompanyProfile">Company profile</a>
+</body>
+</html>
